// app/src/timepicker/utils/types.ts
export declare type OptionTypes = {
// ...
  clockType?: '12h' | '24h';
    /**
     * @description Turn on/off allowNull
     * @default "false"
     */
    allowNull?: boolean;
// ...
  disabledTime?: {
    minutes?: Array<string | number>;
    hours?: Array<string | number>;
    interval?: string;
  };
  /**
   * @description Set current time to the input and timepicker.\
   * If this options is set to `true` it's gonna update picker with toLocaleTimeString() and input with value based on your location.
   * This option also allows to put object with properties.
   * @example
      currentTime: {
        time: new Date(),
        updateInput: true,
        locales: "en-US",
        preventClockType: false
      };
// ...
};